requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic[dotenv]
python-multipart
aiosmtplib
email-validator
cryptography
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
Jinja2
weasyprint
qrcode
pillow